+++ pyspelling/__init__.py
@@ -53,17 +53,6 @@ class SpellChecker(object):
         if self.verbose >= level:
             print(text)
 
-    def _normalize_utf(self, encoding):
-        """Normalize UTF encoding."""
-
-        if encoding == 'utf-8-sig':
-            encoding = 'utf-8'
-        if encoding.startswith('utf-16'):
-            encoding = 'utf-16'
-        elif encoding.startswith('utf-32'):
-            encoding = 'utf-32'
-        return encoding
-
     def setup_command(self, encoding, options, personal_dict):
         """Setup the command."""
 
@@ -95,7 +84,7 @@ class SpellChecker(object):
             if source._has_error():
                 yield util.Results([], source.context, source.category, source.error)
             else:
-                encoding = self._normalize_utf(source.encoding)
+                encoding = source.encoding
                 if source._is_bytes():
                     text = source.text
                 else:
@@ -104,7 +93,7 @@ class SpellChecker(object):
                 cmd = self.setup_command(encoding, options, personal_dict)
                 self.log(str(cmd), 2)
 
-                wordlist = util.call_spellchecker(cmd, input_text=text, encoding=codecs.lookup(encoding).name)
+                wordlist = util.call_spellchecker(cmd, input_text=text, encoding=encoding)
                 yield util.Results([w for w in sorted(set(wordlist.split('\n'))) if w], source.context, source.category)
 
     def compile_dictionary(self, lang, wordlists, output):
@@ -250,7 +239,7 @@ class Aspell(SpellChecker):
         cmd = [
             self.binary,
             'list',
-            '--encoding', codecs.lookup(encoding).name
+            '--encoding', encoding
         ]
 
         if personal_dict:
@@ -342,7 +331,7 @@ class Hunspell(SpellChecker):
         cmd = [
             self.binary,
             '-l',
-            '-i', codecs.lookup(encoding).name
+            '-i', encoding
         ]
 
         if personal_dict: